backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
motor>=2.5.1
pymongo>=3.12.0
python-multipart>=0.0.5
pydantic>=1.8.2
authing-py-sdk>=4.3.3
